+++ packages/ai/src/generate-text/step-result.ts
@@ -165,9 +165,9 @@ export type StepResult<TOOLS extends ToolSet> = {
   readonly providerMetadata: ProviderMetadata | undefined;
 };
 
-export class DefaultStepResult<TOOLS extends ToolSet>
-  implements StepResult<TOOLS>
-{
+export class DefaultStepResult<
+  TOOLS extends ToolSet,
+> implements StepResult<TOOLS> {
   readonly stepNumber: StepResult<TOOLS>['stepNumber'];
   readonly model: StepResult<TOOLS>['model'];
   readonly functionId: StepResult<TOOLS>['functionId'];