+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.88
+
+### Patch Changes
+
+- Updated dependencies [94aed06]
+  - @ai-sdk/anthropic@3.0.65
+
+## 4.0.87
+
+### Patch Changes
+
+- df099b9: feat(bedrock): add support for service tier for model inference
+
+## 4.0.86
+
+### Patch Changes
+
+- 086c3f9: fix(provider/amazon-bedrock): transform bedrock/anthropic error responses to anthropic format
+
+## 4.0.85
+
+### Patch Changes
+
+- 4918dda: fix(amazon-bedrock): preserve reasoning text when signature is present
+
+## 4.0.84
+
+### Patch Changes
+
+- ff854a2: fix(provider/amazon-bedrock): add tool search beta for Anthropic
+
+## 4.0.83
+
+### Patch Changes
+
+- Updated dependencies [05b8ca2]
+  - @ai-sdk/anthropic@3.0.64
+
+## 4.0.82
+
+### Patch Changes
+
+- 055cd68: fix: publish v6 to latest npm dist tag
+- Updated dependencies [055cd68]
+  - @ai-sdk/provider-utils@4.0.21
+  - @ai-sdk/anthropic@3.0.63
+
 ## 4.0.81
 
 ### Patch Changes
@@ -621,13 +668,13 @@
   Before
 
   ```ts
-  model.textEmbeddingModel('my-model-id');
+  model.textEmbeddingModel("my-model-id");
   ```
 
   After
 
   ```ts
-  model.embeddingModel('my-model-id');
+  model.embeddingModel("my-model-id");
   ```
 
 - 2625a04: feat(openai); update spec for mcp approval
@@ -1019,13 +1066,13 @@
   Before
 
   ```ts
-  model.textEmbeddingModel('my-model-id');
+  model.textEmbeddingModel("my-model-id");
   ```
 
   After
 
   ```ts
-  model.embeddingModel('my-model-id');
+  model.embeddingModel("my-model-id");
   ```
 
 - Updated dependencies [8d9e8ad]