+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,90 @@
 
 ### Patch Changes
 
+- ai@6.0.143
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [6f75953]
+  - ai@6.0.142
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.141
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.140
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.139
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.138
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.137
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.136
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [df6a330]
+  - ai@6.0.135
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ed6876b]
+  - ai@6.0.134
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [055cd68]
+  - ai@6.0.133
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [28fd5a5]
+  - ai@6.0.132
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [14f25f9]
+  - ai@6.0.131
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.130
 
 ## 0.0.1